+++ flink-yarn/src/main/java/org/apache/flink/yarn/YarnFlinkApplicationMasterRunner.java
@@ -19,10 +19,15 @@
 package org.apache.flink.yarn;
 
 import akka.actor.ActorSystem;
+import java.util.Map;
+import java.util.concurrent.Callable;
 import org.apache.flink.api.common.JobExecutionResult;
 import org.apache.flink.api.common.time.Time;
 import org.apache.flink.configuration.ConfigConstants;
 import org.apache.flink.configuration.Configuration;
+import org.apache.flink.configuration.GlobalConfiguration;
+import org.apache.flink.configuration.HighAvailabilityOptions;
+import org.apache.flink.configuration.SecurityOptions;
 import org.apache.flink.runtime.akka.AkkaUtils;
 import org.apache.flink.runtime.clusterframework.ApplicationStatus;
 import org.apache.flink.runtime.clusterframework.BootstrapTools;
@@ -42,9 +47,15 @@ import org.apache.flink.runtime.resourcemanager.ResourceManagerRuntimeServicesCo
 import org.apache.flink.runtime.rpc.FatalErrorHandler;
 import org.apache.flink.runtime.rpc.RpcService;
 import org.apache.flink.runtime.rpc.akka.AkkaRpcService;
+import org.apache.flink.runtime.security.SecurityUtils;
 import org.apache.flink.runtime.util.EnvironmentInformation;
 import org.apache.flink.runtime.util.JvmShutdownSafeguard;
 import org.apache.flink.runtime.util.SignalHandler;
+import org.apache.flink.util.Preconditions;
+import org.apache.flink.yarn.cli.FlinkYarnSessionCli;
+import org.apache.hadoop.fs.CommonConfigurationKeysPublic;
+import org.apache.hadoop.security.UserGroupInformation;
+import org.apache.hadoop.yarn.api.ApplicationConstants.Environment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scala.concurrent.duration.FiniteDuration;
@@ -70,8 +81,7 @@ import java.io.ObjectInputStream;
  * JobMaster handles Flink job execution, while the YarnResourceManager handles container
  * allocation and failure detection.
  */
-public class YarnFlinkApplicationMasterRunner extends AbstractYarnFlinkApplicationMasterRunner
-		implements OnCompletionActions, FatalErrorHandler {
+public class YarnFlinkApplicationMasterRunner implements OnCompletionActions, FatalErrorHandler {
 
 	/** Logger */
 	protected static final Logger LOG = LoggerFactory.getLogger(YarnFlinkApplicationMasterRunner.class);
@@ -79,31 +89,52 @@ public class YarnFlinkApplicationMasterRunner extends AbstractYarnFlinkApplicati
 	/** The job graph file path */
 	private static final String JOB_GRAPH_FILE_PATH = "flink.jobgraph.path";
 
+	/** The process environment variables */
+	protected static final Map<String, String> ENV = System.getenv();
+
+	/** The exit code returned if the initialization of the application master failed */
+	protected static final int INIT_ERROR_EXIT_CODE = 31;
+
+	/** The host name passed by env */
+	protected String appMasterHostname;
+
 	// ------------------------------------------------------------------------
 
 	/** The lock to guard startup / shutdown / manipulation methods */
 	private final Object lock = new Object();
 
-	@GuardedBy("lock")
 	private MetricRegistry metricRegistry;
 
-	@GuardedBy("lock")
 	private HighAvailabilityServices haServices;
 
-	@GuardedBy("lock")
 	private HeartbeatServices heartbeatServices;
 
-	@GuardedBy("lock")
 	private RpcService commonRpcService;
 
-	@GuardedBy("lock")
 	private ResourceManager resourceManager;
 
-	@GuardedBy("lock")
 	private JobManagerRunner jobManagerRunner;
 
 	@GuardedBy("lock")
 	private JobGraph jobGraph;
+
+	public YarnFlinkApplicationMasterRunner() {}
+
+	public YarnFlinkApplicationMasterRunner(
+		MetricRegistry metricRegistry,
+		HighAvailabilityServices haServices,
+		HeartbeatServices heartbeatServices,
+		RpcService commonRpcService,
+		ResourceManager resourceManager,
+		JobManagerRunner jobManagerRunner) {
+
+		this.metricRegistry = metricRegistry;
+		this.haServices = haServices;
+		this.heartbeatServices = heartbeatServices;
+		this.commonRpcService = commonRpcService;
+		this.resourceManager = resourceManager;
+		this.jobManagerRunner = jobManagerRunner;
+	}
 
 	// ------------------------------------------------------------------------
 	//  Program entry point
@@ -120,48 +151,176 @@ public class YarnFlinkApplicationMasterRunner extends AbstractYarnFlinkApplicati
 		JvmShutdownSafeguard.installAsShutdownHook(LOG);
 
 		// run and exit with the proper return code
-		int returnCode = new YarnFlinkApplicationMasterRunner().run(args);
+		int returnCode = new YarnFlinkApplicationMasterRunner().initServiceAndComponents().run(args);
 		System.exit(returnCode);
 	}
 
-	@Override
-	protected int runApplicationMaster(Configuration config) {
+	private YarnFlinkApplicationMasterRunner initServiceAndComponents() {
+		LOG.info("Starting High Availability Services");
+		final String currDir = ENV.get(Environment.PWD.key());
+		final String remoteKeytabPath = ENV.get(YarnConfigKeys.KEYTAB_PATH);
+		final String remoteKeytabPrincipal = ENV.get(YarnConfigKeys.KEYTAB_PRINCIPAL);
+
+		// Flink configuration
+		final Map<String, String> dynamicProperties =
+			FlinkYarnSessionCli.getDynamicProperties(ENV.get(YarnConfigKeys.ENV_DYNAMIC_PROPERTIES));
+		final Configuration config = createConfiguration(currDir, dynamicProperties);
+		String keytabPath = null;
+		if(remoteKeytabPath != null) {
+			File f = new File(currDir, Utils.KEYTAB_FILE_NAME);
+			keytabPath = f.getAbsolutePath();
+			LOG.debug("Keytab path: {}", keytabPath);
+		}
+		if (keytabPath != null && remoteKeytabPrincipal != null) {
+			config.setString(SecurityOptions.KERBEROS_LOGIN_KEYTAB, keytabPath);
+			config.setString(SecurityOptions.KERBEROS_LOGIN_PRINCIPAL, remoteKeytabPrincipal);
+		}
+
+		// try to start the rpc service
+		// using the port range definition from the config.
+		final String amPortRange = config.getString(
+			ConfigConstants.YARN_APPLICATION_MASTER_PORT,
+			ConfigConstants.DEFAULT_YARN_JOB_MANAGER_PORT);
 
 		try {
 			// ---- (1) create common services
+			HighAvailabilityServices haServices = HighAvailabilityServicesUtils.createAvailableOrEmbeddedServices(config);
+			HeartbeatServices heartbeatServices = HeartbeatServices.fromConfiguration(config);
+			MetricRegistry metricRegistry = new MetricRegistry(MetricRegistryConfiguration.fromConfiguration(config));
+			RpcService commonRpcService = createRpcService(config, appMasterHostname, amPortRange);
 
-			// try to start the rpc service
-			// using the port range definition from the config.
-			final String amPortRange = config.getString(
-					ConfigConstants.YARN_APPLICATION_MASTER_PORT,
-					ConfigConstants.DEFAULT_YARN_JOB_MANAGER_PORT);
+			// ---- (2) init resource manager -------
+			ResourceManager resourceManager = createResourceManager(config);
 
+			// ---- (3) init job master parameters
+			JobManagerRunner jobManagerRunner = createJobManagerRunner(config);
+
+			return new YarnFlinkApplicationMasterRunner(
+					metricRegistry,
+					haServices,
+					heartbeatServices,
+					commonRpcService,
+					resourceManager,
+					jobManagerRunner);
+		} catch (Exception e) {
+			LOG.error("YARN Application Master initialization failed", e);
+			shutdown(ApplicationStatus.FAILED, e.getMessage());
+			return null;
+		}
+	}
+
+	/**
+	 * The instance entry point for the YARN application master. Obtains user group
+	 * information and calls the main work method {@link #runApplicationMaster(org.apache.flink.configuration.Configuration)} as a
+	 * privileged action.
+	 *
+	 * @param args The command line arguments.
+	 * @return The process exit code.
+	 */
+	protected int run(String[] args) {
+		try {
+			LOG.debug("All environment variables: {}", ENV);
+
+			final String yarnClientUsername = ENV.get(YarnConfigKeys.ENV_HADOOP_USER_NAME);
+			Preconditions.checkArgument(yarnClientUsername != null, "YARN client user name environment variable {} not set",
+				YarnConfigKeys.ENV_HADOOP_USER_NAME);
+
+			final String currDir = ENV.get(Environment.PWD.key());
+			Preconditions.checkArgument(currDir != null, "Current working directory variable (%s) not set", Environment.PWD.key());
+			LOG.debug("Current working directory: {}", currDir);
+
+			final String remoteKeytabPath = ENV.get(YarnConfigKeys.KEYTAB_PATH);
+			LOG.debug("Remote keytab path obtained {}", remoteKeytabPath);
+
+			final String remoteKeytabPrincipal = ENV.get(YarnConfigKeys.KEYTAB_PRINCIPAL);
+			LOG.info("Remote keytab principal obtained {}", remoteKeytabPrincipal);
+
+			String keytabPath = null;
+			if(remoteKeytabPath != null) {
+				File f = new File(currDir, Utils.KEYTAB_FILE_NAME);
+				keytabPath = f.getAbsolutePath();
+				LOG.debug("Keytab path: {}", keytabPath);
+			}
+
+			UserGroupInformation currentUser = UserGroupInformation.getCurrentUser();
+
+			LOG.info("YARN daemon is running as: {} Yarn client user obtainer: {}",
+				currentUser.getShortUserName(), yarnClientUsername );
+
+			// Flink configuration
+			final Map<String, String> dynamicProperties =
+				FlinkYarnSessionCli.getDynamicProperties(ENV.get(YarnConfigKeys.ENV_DYNAMIC_PROPERTIES));
+			LOG.debug("YARN dynamic properties: {}", dynamicProperties);
+
+			final Configuration flinkConfig = createConfiguration(currDir, dynamicProperties);
+			if (keytabPath != null && remoteKeytabPrincipal != null) {
+				flinkConfig.setString(SecurityOptions.KERBEROS_LOGIN_KEYTAB, keytabPath);
+				flinkConfig.setString(SecurityOptions.KERBEROS_LOGIN_PRINCIPAL, remoteKeytabPrincipal);
+			}
+
+			org.apache.hadoop.conf.Configuration hadoopConfiguration = null;
+
+			//To support Yarn Secure Integration Test Scenario
+			File krb5Conf = new File(currDir, Utils.KRB5_FILE_NAME);
+			if (krb5Conf.exists() && krb5Conf.canRead()) {
+				String krb5Path = krb5Conf.getAbsolutePath();
+				LOG.info("KRB5 Conf: {}", krb5Path);
+				hadoopConfiguration = new org.apache.hadoop.conf.Configuration();
+				hadoopConfiguration.set(CommonConfigurationKeysPublic.HADOOP_SECURITY_AUTHENTICATION, "kerberos");
+				hadoopConfiguration.set(CommonConfigurationKeysPublic.HADOOP_SECURITY_AUTHORIZATION, "true");
+			}
+
+			SecurityUtils.SecurityConfiguration sc;
+			if(hadoopConfiguration != null) {
+				sc = new SecurityUtils.SecurityConfiguration(flinkConfig, hadoopConfiguration);
+			} else {
+				sc = new SecurityUtils.SecurityConfiguration(flinkConfig);
+			}
+
+			SecurityUtils.install(sc);
+
+			// Note that we use the "appMasterHostname" given by YARN here, to make sure
+			// we use the hostnames given by YARN consistently throughout akka.
+			// for akka "localhost" and "localhost.localdomain" are different actors.
+			this.appMasterHostname = ENV.get(Environment.NM_HOST.key());
+			Preconditions.checkArgument(appMasterHostname != null,
+				"ApplicationMaster hostname variable %s not set", Environment.NM_HOST.key());
+			LOG.info("YARN assigned hostname for application master: {}", appMasterHostname);
+
+			return SecurityUtils.getInstalledContext().runSecured(new Callable<Integer>() {
+				@Override
+				public Integer call() throws Exception {
+					return runApplicationMaster(flinkConfig);
+				}
+			});
+
+		}
+		catch (Throwable t) {
+			// make sure that everything whatever ends up in the log
+			LOG.error("YARN Application Master initialization failed", t);
+			return INIT_ERROR_EXIT_CODE;
+		}
+	}
+
+	/**
+	 * The main work method, must run as a privileged action.
+	 *
+	 * @return The return code for the Java process.
+	 */
+	protected int runApplicationMaster(Configuration config) {
+
+		try {
 			synchronized (lock) {
-				LOG.info("Starting High Availability Services");
-				haServices = HighAvailabilityServicesUtils.createAvailableOrEmbeddedServices(config);
-
-				heartbeatServices = HeartbeatServices.fromConfiguration(config);
-				
-				metricRegistry = new MetricRegistry(MetricRegistryConfiguration.fromConfiguration(config));
-				commonRpcService = createRpcService(config, appMasterHostname, amPortRange);
-
-				// ---- (2) init resource manager -------
-				resourceManager = createResourceManager(config);
-
-				// ---- (3) init job master parameters
-				jobManagerRunner = createJobManagerRunner(config);
-
-				// ---- (4) start the resource manager  and job manager runner:
+				// ---- (1) start the resource manager  and job manager runner:
 				resourceManager.start();
 				LOG.debug("YARN Flink Resource Manager started");
 
 				jobManagerRunner.start();
 				LOG.debug("Job Manager Runner started");
 
-				// ---- (5) start the web monitor
+				// ---- (2) start the web monitor
 				// TODO: add web monitor
 			}
-
 			// wait for resource manager to finish
 			resourceManager.getTerminationFuture().get();
 			// everything started, we can wait until all is done or the process is killed
@@ -169,8 +328,7 @@ public class YarnFlinkApplicationMasterRunner extends AbstractYarnFlinkApplicati
 		}
 		catch (Throwable t) {
 			// make sure that everything whatever ends up in the log
-			LOG.error("YARN Application Master initialization failed", t);
-			shutdown(ApplicationStatus.FAILED, t.getMessage());
+			LOG.error("Run application master failed", t);
 			return INIT_ERROR_EXIT_CODE;
 		}
 
@@ -180,6 +338,56 @@ public class YarnFlinkApplicationMasterRunner extends AbstractYarnFlinkApplicati
 	// ------------------------------------------------------------------------
 	//  Utilities
 	// ------------------------------------------------------------------------
+
+	/**
+	 * @param baseDirectory  The working directory
+	 * @param additional Additional parameters
+	 *
+	 * @return The configuration to be used by the TaskExecutors.
+	 */
+	private static Configuration createConfiguration(String baseDirectory, Map<String, String> additional) {
+		LOG.info("Loading config from directory {}.", baseDirectory);
+
+		Configuration configuration = GlobalConfiguration.loadConfiguration(baseDirectory);
+
+		// add dynamic properties to JobManager configuration.
+		for (Map.Entry<String, String> property : additional.entrySet()) {
+			configuration.setString(property.getKey(), property.getValue());
+		}
+
+		// override zookeeper namespace with user cli argument (if provided)
+		String cliZKNamespace = ENV.get(YarnConfigKeys.ENV_ZOOKEEPER_NAMESPACE);
+		if (cliZKNamespace != null && !cliZKNamespace.isEmpty()) {
+			configuration.setString(HighAvailabilityOptions.HA_CLUSTER_ID, cliZKNamespace);
+		}
+
+		// if a web monitor shall be started, set the port to random binding
+		if (configuration.getInteger(ConfigConstants.JOB_MANAGER_WEB_PORT_KEY, 0) >= 0) {
+			configuration.setInteger(ConfigConstants.JOB_MANAGER_WEB_PORT_KEY, 0);
+		}
+
+		// if the user has set the deprecated YARN-specific config keys, we add the
+		// corresponding generic config keys instead. that way, later code needs not
+		// deal with deprecated config keys
+
+		BootstrapTools.substituteDeprecatedConfigKey(configuration,
+			ConfigConstants.YARN_HEAP_CUTOFF_RATIO,
+			ConfigConstants.CONTAINERIZED_HEAP_CUTOFF_RATIO);
+
+		BootstrapTools.substituteDeprecatedConfigKey(configuration,
+			ConfigConstants.YARN_HEAP_CUTOFF_MIN,
+			ConfigConstants.CONTAINERIZED_HEAP_CUTOFF_MIN);
+
+		BootstrapTools.substituteDeprecatedConfigPrefix(configuration,
+			ConfigConstants.YARN_APPLICATION_MASTER_ENV_PREFIX,
+			ConfigConstants.CONTAINERIZED_MASTER_ENV_PREFIX);
+
+		BootstrapTools.substituteDeprecatedConfigPrefix(configuration,
+			ConfigConstants.YARN_TASK_MANAGER_ENV_PREFIX,
+			ConfigConstants.CONTAINERIZED_TASK_MANAGER_ENV_PREFIX);
+
+		return configuration;
+	}
 
 	protected RpcService createRpcService(
 			Configuration configuration,
